# kernel-openapi.yaml
openapi: 3.1.0
info:
  title: Institutional Kernel API
  version: "1.1.0"
  description: >
    API HTTP del Institutional Kernel de Mi Rotaract. Fuente única de verdad
    para identidad, organizaciones jerárquicas, membresías, períodos, cargos
    y autoridades, autorización contextual, solicitudes de ingreso,
    transferencias entre organizaciones y registro de módulos externos.
    Este documento es un contrato de datos derivado de `kernel-spec.md`
    (secciones 4, 5, 9 y 10). No reemplaza la especificación; ante cualquier
    conflicto, `kernel-spec.md` es la fuente normativa. El proceso de
    compatibilidad y publicación está definido en
    `kernel-openapi-contract-plan.md`.
  contact:
    name: Institutional Kernel
  license:
    name: Uso interno — Mi Rotaract
    identifier: LicenseRef-MiRotaract-Internal
...
servers:
  - url: https://api.agendai.com.ar/api/kernel/v1
    description: Producción
  - url: https://staging-api.agendai.com.ar/api/kernel/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /persons/{personId}:
    get:
      operationId: getPerson
      tags: [Persons]
      summary: Obtener una persona
      description: >
        Requiere `kernel.person.read`, o `kernel.person.read.self` si
        `personId` corresponde a la persona vinculada a la cuenta autenticada.
      x-required-permission: kernel.person.read
      parameters:
        - $ref: "#/components/parameters/personId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updatePerson
      tags: [Persons]
      summary: Actualizar datos de una persona
      description: >
        Requiere `kernel.person.manage`, o `kernel.person.update.self`
        sobre la propia persona.
      x-required-permission: kernel.person.manage
      parameters:
        - $ref: "#/components/parameters/personId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePersonRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
...
  /organizations:
    post:
      operationId: createOrganization
      tags: [Organizations]
      summary: Crear una organización
      x-required-permission: kernel.organization.create
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrganizationRequest" }
      responses:
        "201":
          description: Creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
    get:
      operationId: listOrganizations
      tags: [Organizations]
      summary: Listar organizaciones
      x-required-permission: kernel.organization.read
      parameters:
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/OrganizationType" }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/OrganizationStatus" }
        - name: parentId
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/query"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationPage" }
  /organizations/{organizationId}:
    get:
      operationId: getOrganization
      tags: [Organizations]
      summary: Obtener una organización
      x-required-permission: kernel.organization.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateOrganization
      tags: [Organizations]
      summary: Actualizar datos de una organización
      x-required-permission: kernel.organization.update
      parameters:
        - $ref: "#/components/parameters/organizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrganizationRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
  /organizations/{organizationId}/activate:
    post:
      operationId: activateOrganization
      tags: [Organizations]
      summary: Activar una organización (DRAFT/INACTIVE → ACTIVE)
      x-required-permission: kernel.organization.activate
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /organizations/{organizationId}/deactivate:
    post:
      operationId: deactivateOrganization
      tags: [Organizations]
      summary: Desactivar una organización (ACTIVE → INACTIVE)
      description: >
        (Supuesto) `kernel-spec.md` §10.1 no define un permiso distinto para
        desactivar; se asume que reutiliza `kernel.organization.activate`
        por tratarse de la misma transición administrativa reversible.
      x-required-permission: kernel.organization.activate
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /organizations/{organizationId}/archive:
    post:
      operationId: archiveOrganization
      tags: [Organizations]
      summary: Archivar una organización (terminal)
      x-required-permission: kernel.organization.archive
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
  /organizations/{organizationId}/move:
    post:
      operationId: moveOrganization
      tags: [Organizations]
      summary: Reasignar el padre de una organización
      description: Valida que el cambio no genere ciclos (invariante 6.3.6).
      x-required-permission: kernel.organization.move
      parameters:
        - $ref: "#/components/parameters/organizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newParentId]
              properties:
                newParentId: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "409":
          description: El movimiento genera un ciclo jerárquico (CA-ORG-03).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
  /organizations/{organizationId}/children:
    get:
      operationId: listOrganizationChildren
      tags: [Organizations]
      summary: Listar hijos directos
      x-required-permission: kernel.organization.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Organization" }
  /organizations/{organizationId}/ancestors:
    get:
      operationId: listOrganizationAncestors
      tags: [Organizations]
      summary: Listar ancestros ordenados desde el padre inmediato
      x-required-permission: kernel.organization.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Organization" }
  /organizations/{organizationId}/descendants:
    get:
      operationId: listOrganizationDescendants
      tags: [Organizations]
      summary: Listar todos los descendientes
      x-required-permission: kernel.organization.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Organization" }

  # ---------------------------------------------------------------------
  # 9.5 Membresías
  # ---------------------------------------------------------------------
...
  /authorization/check:
    post:
      operationId: checkAuthorization
      tags: [Authorization]
      summary: Evaluar una decisión de autorización puntual
      description: >
        Denegación por defecto (principio 9). `DENY` gana sobre `ALLOW` con
        igual o mayor especificidad (invariante 6.7.4).
        Un sujeto autenticado siempre puede consultar su propia
        autorización (`subjectId` = persona propia); consultar la de otro
        `subjectId` requiere `kernel.role.read`.
      x-required-permission: kernel.role.read
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthorizationCheckRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthorizationDecision" }
  /authorization/batch-check:
    post:
      operationId: batchCheckAuthorization
      tags: [Authorization]
      summary: Evaluar hasta 100 decisiones de autorización (CA-AUTHZ-06)
      x-required-permission: kernel.role.read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [checks]
              properties:
                checks:
                  type: array
                  maxItems: 100
                  items:
                    { $ref: "#/components/schemas/AuthorizationCheckRequest" }
      responses:
        "200":
          description: >
            Devuelve una decisión por cada ítem, en el mismo orden
            (CA-AUTHZ-06).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuthorizationDecision" }
...
  /service/users/{accountId}/context:
    get:
      operationId: serviceGetUserContext
      tags: [Service]
      summary: Obtener el UserContext de una cuenta (para servicios)
      security: [{ serviceAuth: [] }]
      x-service-only: true
      parameters:
        - name: accountId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserContext" }
  /service/persons/{personId}:
    get:
      operationId: serviceGetPerson
      tags: [Service]
      summary: Obtener resumen de persona (para servicios)
      security: [{ serviceAuth: [] }]
      x-service-only: true
      parameters:
        - $ref: "#/components/parameters/personId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PersonSummary" }
  /service/organizations/{organizationId}:
    get:
      operationId: serviceGetOrganization
      tags: [Service]
      summary: Obtener resumen de organización (para servicios)
      security: [{ serviceAuth: [] }]
      x-service-only: true
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationSummary" }
  /service/organizations/{organizationId}/membership-snapshot:
    get:
      operationId: serviceGetMembershipSnapshot
      tags: [Service]
      summary: Obtener snapshot de membresías (para quórum, listas, etc.)
      security: [{ serviceAuth: [] }]
      x-service-only: true
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: "#/components/schemas/MembershipStatus" }
        - name: at
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipSnapshot" }
  /service/organizations/{organizationId}/authority-snapshot:
    get:
      operationId: serviceGetAuthoritySnapshot
      tags: [Service]
      summary: Obtener snapshot de autoridades vigentes
      security: [{ serviceAuth: [] }]
      x-service-only: true
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - name: periodId
          in: query
          schema: { type: string }
        - name: at
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthoritySnapshot" }
  /service/organizations/{organizationId}/period-snapshot:
    get:
      operationId: serviceGetPeriodSnapshot
      tags: [Service]
      summary: Obtener snapshot del período vigente
      security: [{ serviceAuth: [] }]
      x-service-only: true
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - name: at
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PeriodSnapshot" }
  /service/authorization/check:
    post:
      operationId: serviceCheckAuthorization
      tags: [Service]
      summary: Evaluar autorización (para servicios, sin sesión de usuario)
      security: [{ serviceAuth: [] }]
      x-service-only: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthorizationCheckRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthorizationDecision" }
  /service/authorization/batch-check:
    post:
      operationId: serviceBatchCheckAuthorization
      tags: [Service]
      summary: Evaluar hasta 100 decisiones (para servicios)
      security: [{ serviceAuth: [] }]
      x-service-only: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [checks]
              properties:
                checks:
                  type: array
                  maxItems: 100
                  items:
                    { $ref: "#/components/schemas/AuthorizationCheckRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuthorizationDecision" }
  /service/modules/{moduleId}/installations/{organizationId}:
    get:
      operationId: serviceGetModuleInstallation
      tags: [Service]
      summary: Obtener el estado de instalación de un módulo (para el propio servicio consumidor)
      security: [{ serviceAuth: [] }]
      x-service-only: true
      parameters:
        - $ref: "#/components/parameters/moduleId"
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleInstallation" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token de usuario emitido por `POST /auth/login` o
        `/auth/refresh` (§14.2). Vida útil de 10 minutos. `aud`:
        `mirotaract-platform`.
    serviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token de servicio a servicio firmado (o mTLS equivalente, §14.3).
        `sub` identifica al servicio emisor. `aud` debe ser
        `institutional-kernel`. Requerido para todos los endpoints bajo
        `/service/*`; el kernel rechaza estas rutas con token de usuario.
  parameters:
    personId:
      name: personId
      in: path
      required: true
      schema: { type: string }
    organizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string }
...
    moduleId:
      name: moduleId
      in: path
      required: true
      schema: { type: string }
...
    query:
      name: query
      in: query
      schema: { type: string }
      description: Búsqueda de texto libre.
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Cursor de paginación opaco devuelto por la página anterior.
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      description: Tamaño de página. Por defecto 25, máximo 100 (§18).
...
  responses:
    NotFound:
      description: Recurso inexistente.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidTransition:
      description: La transición de estado solicitada no es válida desde el estado actual.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      description: Problem Details (RFC 9457).
      required: [type, title, status, code, instance]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: Código estable de error, p. ej. `KERNEL_INVALID_TRANSITION`.
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
      example:
        type: https://api.agendai.com.ar/errors/invalid-transition
        title: Invalid state transition
        status: 409
        code: KERNEL_INVALID_TRANSITION
        detail: ACTIVE membership cannot transition directly to PENDING
        instance: /api/kernel/v1/memberships/mem_123/status
        traceId: trc_123
    PageInfo:
      type: object
      properties:
        nextCursor: { type: string, nullable: true }
        hasMore: { type: boolean }

    # --- Enums (idénticos al esquema Prisma, §5) -----------------------
...
    PlatformRole:
      type: string
      enum: [USER, SUPERADMIN]
    OrganizationType:
      type: string
      enum: [DISTRICT, CLUB, OTHER]
    OrganizationStatus:
      type: string
      enum: [DRAFT, ACTIVE, INACTIVE, ARCHIVED]
    MembershipStatus:
      type: string
      enum: [PENDING, ACTIVE, ON_LEAVE, INACTIVE, GRADUATED, TRANSFERRED]
...
    ScopeType:
      type: string
      enum: [PLATFORM, ORGANIZATION, ORGANIZATION_TREE]
...
    InstallationStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED, DISABLED]
...
    Person:
      type: object
      required: [id, firstName, lastName, createdAt, updatedAt]
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        displayName: { type: string, nullable: true }
        primaryEmail: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true }
        birthDate: { type: string, format: date, nullable: true }
        avatarUrl: { type: string, nullable: true }
        externalReference: { type: string, nullable: true }
        metadata: { type: object, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        archivedAt: { type: string, format: date-time, nullable: true }
    PersonSummary:
      type: object
      description: Vista mínima de Person para consumo entre servicios (§12.5 — sin email salvo necesidad explícita).
      required: [id, firstName, lastName]
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        displayName: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
...
    UpdatePersonRequest:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        displayName: { type: string, nullable: true }
        primaryEmail: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true }
        birthDate: { type: string, format: date, nullable: true }
        avatarUrl: { type: string, nullable: true }
        metadata: { type: object, nullable: true }
    Organization:
      type: object
      required:
        [id, type, code, name, slug, status, timezone, createdAt, updatedAt]
      properties:
        id: { type: string }
        parentId: { type: string, nullable: true }
        type: { $ref: "#/components/schemas/OrganizationType" }
        code: { type: string }
        name: { type: string }
        slug: { type: string }
        status: { $ref: "#/components/schemas/OrganizationStatus" }
        countryCode: { type: string, nullable: true }
        region: { type: string, nullable: true }
        city: { type: string, nullable: true }
        timezone: { type: string, default: "America/Argentina/Cordoba" }
        contactEmail: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        foundedAt: { type: string, format: date, nullable: true }
        description: { type: string, nullable: true }
        attributes: { type: object, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        archivedAt: { type: string, format: date-time, nullable: true }
    OrganizationSummary:
      type: object
      required: [id, type, code, name, slug, status]
      properties:
        id: { type: string }
        parentId: { type: string, nullable: true }
        type: { $ref: "#/components/schemas/OrganizationType" }
        code: { type: string }
        name: { type: string }
        slug: { type: string }
        status: { $ref: "#/components/schemas/OrganizationStatus" }
    OrganizationPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Organization" }
        pageInfo: { $ref: "#/components/schemas/PageInfo" }
    CreateOrganizationRequest:
      type: object
      required: [type, code, name, slug]
      properties:
        parentId: { type: string, nullable: true }
        type: { $ref: "#/components/schemas/OrganizationType" }
        code: { type: string }
        name: { type: string }
        slug: { type: string }
        countryCode: { type: string, nullable: true }
        region: { type: string, nullable: true }
        city: { type: string, nullable: true }
        timezone: { type: string, nullable: true }
        contactEmail: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        foundedAt: { type: string, format: date, nullable: true }
        description: { type: string, nullable: true }
        attributes: { type: object, nullable: true }
    UpdateOrganizationRequest:
      type: object
      properties:
        name: { type: string }
        countryCode: { type: string, nullable: true }
        region: { type: string, nullable: true }
        city: { type: string, nullable: true }
        timezone: { type: string }
        contactEmail: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        description: { type: string, nullable: true }
        attributes: { type: object, nullable: true }
...
    AuthorizationCheckRequest:
      type: object
      required: [subjectId, permission, scope]
      properties:
        subjectId:
          type: string
          description: personId a evaluar.
        permission:
          type: string
          example: meetings.meeting.create
        scope:
          type: object
          required: [type]
          properties:
            type: { $ref: "#/components/schemas/ScopeType" }
            organizationId: { type: string, nullable: true }
        periodId: { type: string, nullable: true }
        resource:
          type: object
          nullable: true
          properties:
            type: { type: string }
            id: { type: string }
            attributes: { type: object }
    AuthorizationDecision:
      type: object
      required: [allowed, decisionId, subjectId, permission, evaluatedAt]
      properties:
        allowed: { type: boolean }
        decisionId: { type: string }
        subjectId: { type: string }
        permission: { type: string }
        matchedAssignments:
          type: array
          items: { type: string }
        reasonCodes:
          type: array
          items: { type: string }
          example: ["ROLE_ALLOWED"]
        evaluatedAt: { type: string, format: date-time }
        cacheUntil: { type: string, format: date-time, nullable: true }
...
    ModuleInstallation:
      type: object
      required: [id, moduleId, organizationId, status, installedAt, updatedAt]
      properties:
        id: { type: string }
        moduleId: { type: string }
        organizationId: { type: string }
        status: { $ref: "#/components/schemas/InstallationStatus" }
        configuration: { type: object, nullable: true }
        installedById: { type: string, nullable: true }
        installedAt: { type: string, format: date-time }
        activatedAt: { type: string, format: date-time, nullable: true }
        disabledAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time }

    # --- Contratos para consumidores (§12) ------------------------------
    UserContext:
      type: object
      required:
        [
          accountId,
          personId,
          accountStatus,
          platformRole,
          displayName,
          memberships,
          contextVersion,
        ]
      properties:
        accountId: { type: string }
        personId: { type: string }
        accountStatus: { type: string, enum: [ACTIVE] }
        platformRole: { $ref: "#/components/schemas/PlatformRole" }
        displayName: { type: string }
        memberships:
          type: array
          items:
            type: object
            required: [membershipId, organizationId, organizationType, status]
            properties:
              membershipId: { type: string }
              organizationId: { type: string }
              organizationType:
                { $ref: "#/components/schemas/OrganizationType" }
              status: { type: string }
        contextVersion: { type: integer }
    MembershipSnapshot:
      type: object
      required: [snapshotId, organizationId, capturedAt, sourceVersion, members]
      properties:
        snapshotId: { type: string }
        organizationId: { type: string }
        capturedAt: { type: string, format: date-time }
        sourceVersion: { type: integer }
        members:
          type: array
          items:
            type: object
            required: [membershipId, personId, status]
            properties:
              membershipId: { type: string }
              personId: { type: string }
              accountId: { type: string, nullable: true }
              status: { type: string }
    AuthoritySnapshot:
      type: object
      required: [snapshotId, organizationId, periodId, capturedAt, appointments]
      properties:
        snapshotId: { type: string }
        organizationId: { type: string }
        periodId: { type: string }
        capturedAt: { type: string, format: date-time }
        appointments:
          type: array
          items:
            type: object
            required:
              [
                appointmentId,
                positionCode,
                membershipId,
                membershipOrganizationId,
                personId,
                status,
              ]
            properties:
              appointmentId: { type: string }
              positionCode: { type: string }
              membershipId: { type: string }
              membershipOrganizationId: { type: string }
              personId: { type: string }
              status: { type: string, enum: [ACTIVE] }
              startsAt: { type: string, format: date-time, nullable: true }
              endsAt: { type: string, format: date-time, nullable: true }
    PeriodSnapshot:
      type: object
      description: Snapshot sellado del período activo, o null si no existe.
      required: [snapshotId, organizationId, capturedAt]
      properties:
        snapshotId: { type: string }
        organizationId: { type: string }
        capturedAt: { type: string, format: date-time }
        currentPeriod:
          nullable: true
          type: object
          required: [periodId, code, name, sequence, startDate, endDate, status]
          properties:
            periodId: { type: string }
            code: { type: string }
            name: { type: string }
            sequence: { type: integer }
            startDate: { type: string, format: date }
            endDate: { type: string, format: date }
            status: { type: string, enum: [ACTIVE] }